+++ pyhidra/__init__.py
@@ -1,5 +1,5 @@
 
-__version__ = "1.0.2"
+__version__ = "1.1.0"
 
 # Expose API
 from .core import run_script, start, started, open_program